T5 - CS Solo Online - Quoc Huy/First-AI-Agent/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tabulate
mysql-connector-python